Refactor: extract inline tests to integration test suite
Move all #[cfg(test)] modules from source files into dedicated
integration tests under tests/. Restructure as lib+bin crate so
integration tests can access the library's public API directly.
