+++ src/app/alerts.rs
@@ -171,54 +171,3 @@ impl App {
         triggered
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn check_alerts_fires_when_price_matches() {
-        let mut app = App::new().unwrap();
-        let alert = Alert::new("BBCA", AlertType::Above, 8000.0);
-        app.config.add_alert(alert);
-
-        // Simulate a quote
-        let quote = crate::api::StockQuote {
-            symbol: "BBCA".to_string(),
-            short_name: "Bank Mandiri".to_string(),
-            price: 8001.0,
-            change: 100.0,
-            change_percent: 1.0,
-            open: 7900.0,
-            high: 8050.0,
-            low: 7900.0,
-            volume: 1_000_000,
-            prev_close: 7900.0,
-            long_name: Some("PT Bank Mandiri".to_string()),
-            sector: Some("Financial".to_string()),
-            industry: Some("Banking".to_string()),
-            market_cap: Some(100_000_000_000),
-            trailing_pe: Some(10.0),
-            dividend_yield: Some(2.5),
-            fifty_two_week_high: Some(9000.0),
-            fifty_two_week_low: Some(7000.0),
-            beta: Some(1.2),
-            average_volume: Some(500_000),
-        };
-        app.quotes.insert("BBCA".to_string(), quote);
-
-        let triggered = app.check_alerts();
-        assert_eq!(triggered.len(), 1);
-        assert!(triggered[0].1.contains("crossed above"));
-    }
-
-    #[test]
-    fn open_alert_modal_returns_to_normal_when_no_symbol() {
-        let mut app = App::new().unwrap();
-        app.view_mode = ViewMode::Watchlist;
-        app.selected_index = 999; // Out of bounds
-        app.open_alert_modal();
-        assert_eq!(app.input_mode, InputMode::Normal);
-        assert_eq!(app.alert_symbol, None);
-    }
-}
