+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod api;
+pub mod app;
+pub mod config;
+pub mod ui;
